feat: Add 'account' app to INSTALLED_APPS
This commit adds the 'account' app to the INSTALLED_APPS list in the settings.py file. The 'account' app is a new addition to the project and will provide functionality related to user accounts and authentication. This enhancement ensures that the 'account' app is properly configured and ready to be used in the e-commerce website.

Co-authored-by: Habibur Rahaman Fahim