feat: Add email validation to CreateUserForm
This commit adds email validation to the `CreateUserForm` in the `forms.py` file of the `account` app. The `clean_email` method checks if the entered email already exists in the database and raises a validation error if it does. It also checks if the email length exceeds the maximum allowed limit and raises a validation error in that case. This enhancement ensures that the email input is validated properly during user registration.

Co-authored-by: Habibur Rahaman Fahim