feat: Update registration form layout in register.html
This commit updates the registration form layout in the register.html template of the account app. The changes include adding a CSS class "form-layout" to the container div, which applies a border-radius of 20px and sets the width to 500px. This enhancement improves the visual appearance of the registration form, providing a more modern and consistent design.

Co-authored-by: Habibur Rahaman Fahim